raftstore: split store module (#3528)
* raftstore: restruct store module

This PR divides store module into two parts: one is for overall store
operations, the other is for region related operations. It's very
helpful for threaded raftstore experiment, as it can reduce the conflict
a lot.
