+++ src/raftstore/store/mod.rs
@@ -15,9 +15,9 @@ pub mod bootstrap;
 pub mod cmd_resp;
 pub mod config;
 pub mod engine;
+pub mod fsm;
 pub mod keys;
 pub mod msg;
-pub mod store;
 pub mod transport;
 pub mod util;
 
@@ -35,6 +35,10 @@ pub use self::bootstrap::{
 };
 pub use self::config::Config;
 pub use self::engine::{Iterable, Mutable, Peekable};
+pub use self::fsm::{
+    create_event_loop, new_compaction_listener, DestroyPeerJob, Store, StoreChannel, StoreInfo,
+    StoreStat,
+};
 pub use self::msg::{
     BatchReadCallback, Callback, Msg, ReadCallback, ReadResponse, SeekRegionCallback,
     SeekRegionFilter, SeekRegionResult, SignificantMsg, Tick, WriteCallback, WriteResponse,
@@ -52,7 +56,6 @@ pub use self::snap::{
     check_abort, copy_snapshot, ApplyOptions, Error as SnapError, SnapEntry, SnapKey, SnapManager,
     SnapManagerBuilder, Snapshot, SnapshotDeleter, SnapshotStatistics,
 };
-pub use self::store::{create_event_loop, new_compaction_listener, Store, StoreChannel, StoreStat};
 pub use self::transport::Transport;
 pub use self::util::Engines;
 pub use self::worker::{KeyEntry, ReadTask};
